replace old cris import with new cris import
