+++ atd-etl/cris_import/.dockerignore
@@ -1,1 +1,4 @@
 env
+*.env
+**.pyc
+extracts/*
