+++ atd-etl/cris_import/.gitignore
@@ -1,1 +1,2 @@
-env
+extracts/*
+!extracts/*.md
