--- atd-etl/cris_import/development_extracts/.gitignore
@@ -1,2 +1,0 @@
-*.zip
-
