--- atd-etl/cris_import/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-/usr/sbin/sshd
-/app/cris_import.py
