--- atd-etl/cris_import/env-template
@@ -1,3 +1,0 @@
-OP_API_TOKEN=
-OP_CONNECT=
-OP_VAULT_ID=
