--- atd-etl/cris_import/lib/sshkeytempdir.py
@@ -1,28 +1,0 @@
-import tempfile
-import os
-import shutil
-
-
-# these temp directories are used to store ssh keys, because they will
-# automatically clean themselves up when they go out of scope.
-class SshKeyTempDir:
-    def __init__(self):
-        self.path = None
-
-    def __enter__(self):
-        self.path = tempfile.mkdtemp(dir="/tmp")
-        return self.path
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        shutil.rmtree(self.path)
-
-
-def write_key_to_file(path, content):
-    # Open the file with write permissions and create it if it doesn't exist
-    fd = os.open(path, os.O_WRONLY | os.O_CREAT, 0o600)
-
-    # Write the content to the file
-    os.write(fd, content.encode())
-
-    # Close the file
-    os.close(fd)
